Move the box
