+++ frontend/src/pages/Worksheet.jsx
@@ -30,35 +30,6 @@ const StyledCalendarContainer = styled(SurfaceComponent)`
     color ${({ theme }) => theme.trans_dur};
 `;
 
-function WorksheetStats() {
-  const { courses, hidden_courses, cur_season } = useWorksheet();
-  let courseCnt = 0,
-    credits = 0,
-    workload = 0,
-    rating = 0;
-  for (const c of courses) {
-    if (hidden_courses[cur_season]?.[c.crn]) continue; // just skip hidden courses
-
-    courseCnt += 1;
-    credits += c.credits ?? 0;
-    workload += c.average_workload ?? 0;
-    rating += c.average_rating ?? 0;
-  }
-
-  const avgRating = courseCnt === 0 ? 0 : (rating / courseCnt).toFixed(2); // better to calculate not in the jsx
-
-  return (
-    <div className={styles.stats}>
-      <ul>
-        <li>Total courses: {courseCnt}</li>
-        <li>Total credits: {credits}</li>
-        <li>Total workload: {workload.toFixed(2)}</li>
-        <li>Average rating: {avgRating}</li>
-      </ul>
-    </div>
-  );
-}
-
 /**
  * Renders worksheet page
  */
@@ -164,7 +135,6 @@ function Worksheet() {
                 className={styles.calendar_style_container}
               >
                 <WorksheetCalendar />
-                <WorksheetStats />
                 {/* Expand/Compress icons for calendar */}
                 <StyledExpandBtn
                   className={`${styles.expand_btn} ${styles.top_right}`}
